refactor: Option 타입 분리 selecedOptionProvider.tsx Option interface 분리
