+++ src/@types/Response.ts
@@ -1,6 +1,7 @@
 export interface Option {
   key: string;
-  value: string;
+  value?: string;
+  name?: string;
   category: string;
   price: number;
   id: number;
